Added Problem File and Stubbed Menu Options
Created a text file with all the problems listed so I don't have to access
the website to do the problems. There are hashed solutions in the text
file too, might write a little side project to give the solutions to the
different problems.

Also added a couple of menu options for the next few problems.