Testing and Linked List Stack Implementation
Created stack.c/h which is a simple linked list stack implementation. It's
incomplete at the moment and needs work, but it works.

Also created a quick test for this implementation, which just tries to
push and pop some chars.

This also meant setting up a test dir and amending the makefile to have
test builds and clean rules.

Pushed a lot into this commit, but it all seems to be working well. Famous
last words...