File pointer for Problem 8 now working
Was having issues with the file pointer for reading the series of digits
from file for Problem 8. Issue was with the path I was passing. I was
giving a relative path, but from the wrong directory. The path either
needs to be relative to the dir you're in when you run the exe, or
absolute. I've amended it to be absolute.

There's a couple of bits and pieces I want to move into a separate source
for file handling, checking for access/existence etc.