Reworked problem 3 #1
Added parameter for the number to find greatest prime factor of, for
problem 3. Had to create a new input function to handle possibly reading a
long long int, as the project euler problem does. So there's a new
function get_input_ll() for reading a long long int from user.