Amended is_prime and added test function
Created a test function for is_prime and check_answer functions.

Re implemented is_prime using some shortcuts I found on SO post;
http://stackoverflow.com/questions/1538644/c-determine-if-a-number-is-prime

But now I'm not passing test 7, on the 10001st prime number. I'm not 100%
where the error is happening at the moment, but I think the is_prime
function must be returned around a 100 non-primes on the way to finding
the 10001st, as I end up returning the 9918th...

More testing.