Testing sum_multiples #8
Had a bit of an issue getting 2 of the tests to pass before I realised it
was my maths that was out. I was adding common multiples of each number to
the total.

I changed the sum_multiples function to add both, which ended up failing
the example and actual PE problems, so I'm pretty sure the problem is
meant to only include a single instance of each common multiple.

Amended the expected test results to suit and all tests are passing.