Implemented testing for sum_even_fibonacci_num #17
Changed the sum_even_fibonacci function. The issue was the values for
prev, pprev and curr were updated at the beginning of the loop, before the
sum was added to (if the number was even).

This meant that the condition for the while loop (curr < limit) was
checked, then the next number in the sequence evaluated, then tested for
eveness and added to the sum.

So in the case of the limit being 100, the value of curr=89 passed the
while loop condition, but was then reset to prev+pprev, which was 144,
then this value  was added to the sum as it is even.

So effectively the loop was always going one number in the sequence beyond
the specified limit.

The change was to move the updating of the prev, pprev and curr values to
the end of the loop, after the sum was added to.